+++ lib/bonny/config.ex
@@ -131,10 +131,28 @@ defmodule Bonny.Config do
   end
 
   @doc """
-  `K8s.Cluster` name used for this operator. Defaults to `:default`
+  `K8s.Conn` name used for this operator.
   """
-  @spec cluster_name() :: atom
-  def cluster_name() do
-    Application.get_env(:bonny, :cluster_name, :default)
+  @spec conn() :: K8s.Conn.t()
+  def conn() do
+    get_conn = Application.get_env(:bonny, :get_conn)
+
+    case apply_get_conn(get_conn) do
+      {:ok, %K8s.Conn{} = conn} ->
+        conn
+
+      %K8s.Conn{} = conn ->
+        conn
+
+      _ ->
+        raise("""
+        Check bonny.get_conn in your config.exs. get_conn must be a tuple in the form {Module, :function, [args]}
+        which defines a function returning {:ok, K8s.Conn.t()}. Given: #{inspect(get_conn)}
+        """)
+    end
   end
+
+  defp apply_get_conn({module, function, args}), do: apply(module, function, args)
+  defp apply_get_conn({module, function}), do: apply(module, function, [])
+  defp apply_get_conn(_), do: :error
 end