+++ test/support/k8s_mock.ex
@@ -1,0 +1,13 @@
+# credo:disable-for-this-file
+defmodule Bonny.K8sMock do
+  def conn(mock) do
+    #  Have to register the mock here because Bonny uses Task.start() so we won't know the pid when starting the test.
+    K8s.Client.DynamicHTTPProvider.register(self(), mock)
+
+    %K8s.Conn{
+      discovery_driver: K8s.Discovery.Driver.File,
+      discovery_opts: [config: "test/support/discovery/tests.json"],
+      http_provider: K8s.Client.DynamicHTTPProvider
+    }
+  end
+end