+++ test/test_helper.exs
@@ -1,5 +1,7 @@
+{:ok, _} = Application.ensure_all_started(:k8s)
+{:ok, _} = K8s.Client.DynamicHTTPProvider.start_link(nil)
+
 K8s.Sys.Logger.attach()
 Bonny.Sys.Logger.attach()
 
-{:ok, _} = Whizbang.start_link()
 ExUnit.start()